refactor(workflow-tasks): move verification component into library
